+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/parties/legacy/LegacyGetPartyTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.parties
+package com.forgerock.uk.openbanking.tests.functional.account.parties.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEqualTo
@@ -10,13 +10,12 @@ import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.*
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READACCOUNTSDETAIL
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READPARTYPSU
 
-class GetPartyTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetPartyTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
@@ -99,52 +98,6 @@ class GetPartyTest(val tppResource: CreateTppCallback.TppResource) {
         // when
         val party =
             AccountRS().getAccountsData<OBReadParty2>(accountAndTransaction3_1_2.Links.links.GetParty, accessToken)
-
-        // Then
-        assertThat(party).isNotNull()
-        assertThat(party.data.party).isNotNull()
-    }
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetParty"],
-        apis = ["party"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5", "v.3.1.4", "v.3.1.3"]
-    )
-    @Test
-    fun shouldGetParty_v3_1_8() {
-        // Given
-        val consentRequest = OBReadConsent1().data(
-            OBReadData1()
-                .permissions(listOf(READPARTYPSU, READACCOUNTSDETAIL, OBExternalPermissions1Code.READPARTY))
-        )
-            .risk(OBRisk2())
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-
-        val (_, psuId) = AccountRS().getFirstAccountIdAndPsuId(
-            accountAndTransaction3_1_8.Links.links.GetAccounts,
-            accessToken
-        )
-        assertThat(psuId).isNotEqualTo("")
-        psu.user.uid = psuId
-
-        // when
-        val party =
-            AccountRS().getAccountsDataEndUser<OBReadParty2>(
-                accountAndTransaction3_1_8.Links.links.GetParty,
-                accessToken
-            )
 
         // Then
         assertThat(party).isNotNull()
